Fixed FileList SelectAll button margin
